+++ src/lib.rs
@@ -1,9 +1,13 @@
 mod cast;
+mod containing_float;
+mod goniometric;
 mod isqrt;
 mod sqrt;
 mod vec2;
 
-pub use self::{cast::*, isqrt::*, sqrt::*, vec2::*};
+pub use self::{
+    cast::*, containing_float::*, goniometric::*, isqrt::*, sqrt::*, vec2::*,
+};
 
 #[cfg(test)]
 mod tests {}