+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
 
+## future
+- Add new methods for `Vec2`: `pos_of_idx` and `angle`.
+- Add new traits: `ContainingFloat` and `Goniometric`.
+
 ## v0.1.0
 - Add `Vec2`.